requirements.txt
appdirs==1.4.4
APScheduler==3.6.3
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
cryptography==3.2.1
decorator==4.4.2
distlib==0.3.1
filelock==3.0.12
idna==2.10
loguru==0.5.3
pafy==0.5.5
pycparser==2.20
python-dateutil==2.8.1
python-dotenv==0.15.0
python-telegram-bot==13.0
pytz==2020.1
requests==2.24.0
six==1.15.0
tinydb==4.2.0
tornado==6.1
tzlocal==2.1
urllib3==1.25.11
virtualenv==20.2.1
youtube-dl==2020.12.2
